Add files via upload
My code takes in an array and gets the least amount of values to be greater than the sum of the remaining values. It tests every value of an array with every number and gets the sum of each iteration. Whenever the value is greater than the sum of the remaining values the program breaks and tests if the number of values that is needed to be greater than the remaining values is greater than the previous number of values. The program returns the least amount of values needed from the iterations to be greater than the sum of the remaining values.